requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
openai>=1.13.3
python-dotenv==1.0.0
gunicorn==21.2.0
eventlet==0.33.3
